+++ numpy/fft/__init__.py
@@ -200,12 +200,15 @@ For examples, see the various functions.
 
 """
 
-from . import _pocketfft, helper
+from . import _pocketfft, _helper
+# TODO: `numpy.fft.helper`` was depreacted in NumPy 2.0. It should
+# be deleted once downstream libraries move to `numpy.fft`.
+from . import helper
 from ._pocketfft import *
-from .helper import *
+from ._helper import *
 
 __all__ = _pocketfft.__all__.copy()
-__all__ += helper.__all__
+__all__ += _helper.__all__
 
 from numpy._pytesttester import PytestTester
 test = PytestTester(__name__)
