+++ numpy/fft/__init__.pyi
@@ -17,7 +17,7 @@ from numpy.fft._pocketfft import (
     ifftn as ifftn,
 )
 
-from numpy.fft.helper import (
+from numpy.fft._helper import (
     fftshift as fftshift,
     ifftshift as ifftshift,
     fftfreq as fftfreq,
